moved run_nougat to sexample.py
